+++ js/packages/openinference-instrumentation-beeai/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @arizeai/openinference-instrumentation-beeai
 
+## 1.5.20
+
+### Patch Changes
+
+- Updated dependencies [0168198]
+  - @arizeai/openinference-core@2.5.0
+
+## 1.5.19
+
+### Patch Changes
+
+- Updated dependencies [145e3c6]
+  - @arizeai/openinference-semantic-conventions@2.6.0
+  - @arizeai/openinference-core@2.4.1
+
 ## 1.5.18
 
 ### Patch Changes