+++ js/packages/openinference-instrumentation-openai/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @arizeai/openinference-instrumentation-openai
 
+## 4.1.7
+
+### Patch Changes
+
+- Updated dependencies [0168198]
+  - @arizeai/openinference-core@2.5.0
+
+## 4.1.6
+
+### Patch Changes
+
+- Updated dependencies [145e3c6]
+  - @arizeai/openinference-semantic-conventions@2.6.0
+  - @arizeai/openinference-core@2.4.1
+
 ## 4.1.5
 
 ### Patch Changes